feat(ui): Wave 5 — UX polish (sidebar reorder + BSS icon + marketplace as SettingsCard) + chart 1.4.155
Founder UX-polish review (2026-05-17, post Wave-2 collector). Three
distinct fixes the founder flagged:

1. Sidebar order followed no logic — random walk Apps/Jobs/Dashboard/
   Cloud/Users/BSS. Reordered to operator mental model:
   Dashboard → Cloud → Apps → Jobs → Users → BSS → Settings

2. BSS icon was a bespoke receipt glyph that didn't match the line-
   glyph family. Swapped to a briefcase glyph fitting stylistically.

3. Marketplace toggle was a dedicated /settings/marketplace page +
   Settings sub-nav child. Founder: "if market place is just a toggle
   ... it should be ... similar to other setting". Refactored into
   SettingsPage SectionCard anchor (id=marketplace, same as #dns).
   MarketplaceSettings.tsx + .test.tsx + route + sub-nav child deleted.
   Save flow unchanged: POSTs /api/v1/sovereigns/{id}/marketplace.

Chart 1.4.154 → 1.4.155 + bootstrap-kit pin bump per the
chart-bump-needs-both-files rule.
